requirements.txt
coreapi
Django
django-filter
django-heroku
django-jet
django-recurrence
djangorestframework
djangorestframework-jwt
gunicorn
Markdown
psycopg2
django-storages
Pillow
boto3
